part A and B done
part A and B are done, change the shapes of all the tanks to circles and
use the stuff in part B to make part C.
Look up "Plane.GetSide" that will help a LOT.
Just make two planes that are perpendicular to eachother and the floor.
that will give us a perfect X around our character. finally we're gonna
have 4 public counters. one for each side. then just do the getside
function thing on all the tanks to see what side they're on. increment
the counter for that portion of the X then we're done!